wip created HashVector in collectors/hashvec.py moved bb hashing logic to HashVector made FnvHash and BernsteinHash to inherit from HashVector maybe broke some code..do not merge that yet
